+++ resources/js/Pages/Auth/Layouts/Layout.vue
@@ -1,0 +1,3 @@
+<template>
+    <slot></slot>
+</template>
