+++ resources/js/Pages/Layouts/HomePageLayout.vue
@@ -1,0 +1,3 @@
+<template>
+    <slot></slot>
+</template>
