+++ resources/js/Pages/Main-Platform/Shared/LeftSideNavigation.vue
@@ -1,0 +1,6 @@
+<template>
+    <nav class="flex flex-col gap-5 p-3">
+        <Link>Menu Item 1</Link>
+        <Link>Menu Item 2</Link>
+    </nav>
+</template>
