+++ resources/js/app.js
@@ -1,11 +1,14 @@
 import { createApp, h } from 'vue';
 import { createInertiaApp } from '@inertiajs/vue3';
 import { Link } from '@inertiajs/vue3';
+import DashboardLayout from '@/Pages/Main-Platform/Layouts/DashboardLayout.vue';
 
 createInertiaApp({
     resolve: name => {
         const pages = import.meta.glob('./Pages/**/*.vue', { eager: true });
-        return pages[`./Pages/${name}.vue`];
+        const page = pages[`./Pages/${name}.vue`]
+        page.default.layout = page.default.layout || DashboardLayout
+        return page
     },
     setup({ el, App, props, plugin }) {
         createApp({ render: () => h(App, props) })
